chore(util-credentials): refactor the functions to own files
