+++ packages/util-credentials/src/parse-known-profiles.ts
@@ -1,0 +1,37 @@
+import {
+  loadSharedConfigFiles,
+  ParsedIniData,
+  SharedConfigFiles,
+  SharedConfigInit,
+} from "@aws-sdk/shared-ini-file-loader";
+
+export interface SourceProfileInit extends SharedConfigInit {
+  /**
+   * The configuration profile to use.
+   */
+  profile?: string;
+
+  /**
+   * A promise that will be resolved with loaded and parsed credentials files.
+   * Used to avoid loading shared config files multiple times.
+   *
+   * @internal
+   */
+  loadedConfig?: Promise<SharedConfigFiles>;
+}
+
+/**
+ * Load profiles from credentials and config INI files and normalize them into a
+ * single profile list.
+ *
+ * @internal
+ */
+export const parseKnownFiles = async (init: SourceProfileInit): Promise<ParsedIniData> => {
+  const { loadedConfig = loadSharedConfigFiles(init) } = init;
+
+  const parsedFiles = await loadedConfig;
+  return {
+    ...parsedFiles.configFile,
+    ...parsedFiles.credentialsFile,
+  };
+};
